requirements.txt
requests==2.23.0
python-ev3dev2==2.1.0
